fix build and eslint + fix cumulative usage inflating turns in openrouter agent
