+++ web_src/src/lib/agentRunTelemetry.ts
@@ -145,7 +145,7 @@ function upsertTurn(
       ...(message?.trim() ? { message: message.trim() } : {}),
     },
   ];
-  return next.toSorted((a, b) => a.turn - b.turn);
+  return [...next].sort((a, b) => a.turn - b.turn);
 }
 
 function turnForTool(state: AgentRunTelemetry, recordTurn: number | undefined): number {
@@ -347,46 +347,6 @@ export function usageDelta(current: AgentTurnUsage, previous?: AgentTurnUsage): 
   return delta;
 }
 
-export function tokenTotal(usage: AgentTurnUsage): number {
-  return (
-    usage.input_tokens +
-    usage.output_tokens +
-    usage.cache_read_input_tokens +
-    usage.cache_creation_input_tokens +
-    usage.reasoning_tokens
-  );
-}
-
-export function billedTokenTotal(usage: AgentTurnUsage): number {
-  return usage.input_tokens + usage.output_tokens + usage.cache_creation_input_tokens;
-}
-
-export function inputTokenTotal(usage: AgentTurnUsage): number {
-  return usage.input_tokens + usage.cache_creation_input_tokens;
-}
-
-export function outputTokenTotal(usage: AgentTurnUsage): number {
-  return usage.output_tokens + usage.reasoning_tokens;
-}
-
-export function newTokenTotal(usage: AgentTurnUsage): number {
-  return inputTokenTotal(usage) + outputTokenTotal(usage);
-}
-
-export function usageEquals(left: AgentTurnUsage, right: AgentTurnUsage): boolean {
-  return (
-    left.input_tokens === right.input_tokens &&
-    left.output_tokens === right.output_tokens &&
-    left.cache_read_input_tokens === right.cache_read_input_tokens &&
-    left.cache_creation_input_tokens === right.cache_creation_input_tokens &&
-    left.reasoning_tokens === right.reasoning_tokens
-  );
-}
-
-export function agentRunToolCallCount(telemetry: AgentRunTelemetry): number {
-  return Object.values(telemetry.tool_counts).reduce((sum, count) => sum + count, 0);
-}
-
 export function parseAgentTurnLiveLogText(
   text: string,
 ): { turn: number; usage: AgentTurnUsage; message?: string } | null {
@@ -408,65 +368,6 @@ export function parseAgentTurnLiveLogText(
 
 export function isRawAgentTurnLiveLogText(text: string): boolean {
   return parseAgentTurnLiveLogText(text) != null;
-}
-
-export type AgentChartPoint = {
-  turn: number;
-  tokens: number;
-  input_tokens: number;
-  output_tokens: number;
-  cache_read_tokens: number;
-  tools: number;
-  turnTools: AgentTurnTool[];
-  message: string;
-  inputBar: number;
-  outputBar: number;
-};
-
-export function chartPointsForTelemetry(telemetry: AgentRunTelemetry): AgentChartPoint[] {
-  return telemetry.turns.map((snapshot) => {
-    const usage = snapshot.usage;
-    const inputBar = inputTokenTotal(usage);
-    const outputBar = outputTokenTotal(usage);
-    return {
-      turn: snapshot.turn,
-      tokens: inputBar + outputBar,
-      input_tokens: usage.input_tokens,
-      output_tokens: usage.output_tokens,
-      cache_read_tokens: usage.cache_read_input_tokens,
-      tools: snapshot.tools.length,
-      turnTools: snapshot.tools,
-      message: snapshot.message ?? "",
-      inputBar,
-      outputBar,
-    };
-  });
-}
-
-export function agentRunBilledTokenCount(telemetry: AgentRunTelemetry): number {
-  return agentRunNewTokenCount(telemetry);
-}
-
-export function agentRunNewTokenCount(telemetry: AgentRunTelemetry): number {
-  return agentRunInputTokenCount(telemetry) + agentRunOutputTokenCount(telemetry);
-}
-
-export function agentRunInputTokenCount(telemetry: AgentRunTelemetry): number {
-  return telemetry.turns.reduce((sum, turn) => sum + inputTokenTotal(turn.usage), 0);
-}
-
-export function agentRunOutputTokenCount(telemetry: AgentRunTelemetry): number {
-  return telemetry.turns.reduce((sum, turn) => sum + outputTokenTotal(turn.usage), 0);
-}
-
-export function agentRunCacheReadCount(telemetry: AgentRunTelemetry): number {
-  return telemetry.turns.reduce((sum, turn) => sum + turn.usage.cache_read_input_tokens, 0);
-}
-
-export function agentRunTokenCount(telemetry: AgentRunTelemetry): number {
-  const fromTurns = telemetry.turns.reduce((sum, turn) => sum + tokenTotal(turn.usage), 0);
-  const fromUsage = tokenTotal(telemetry.usage);
-  return Math.max(fromTurns, fromUsage);
 }
 
 export type AgentPromptUsageSeries = {
