+++ web_src/src/pages/factories/pages/work-order-split-run/PhaseLogCard.titleLine.spec.tsx
@@ -8,7 +8,6 @@ import type { FactoriesWorkOrderArtifact } from "@/api-client";
 
 import { PhaseLogCard } from "./PhaseLogCard";
 import { idleLiveLogStream, line, PHASE } from "./PhaseLogCard.testHelpers";
-import type { SplitRunStreamLine } from "./splitRunMocks";
 
 const PLAN_MD_ARTIFACT: FactoriesWorkOrderArtifact = {
   id: "art-plan",
@@ -478,274 +477,5 @@ describe("PhaseLogCard phase actions", () => {
     expect(screen.queryByRole("link", { name: "View automation run" })).not.toBeInTheDocument();
     expect(screen.queryByRole("link", { name: "Edit automation" })).not.toBeInTheDocument();
     expect(screen.queryByRole("button", { name: "Show usage" })).not.toBeInTheDocument();
-  });
-});
-
-function usageTelemetry(inputTokens: number, tools: Array<{ kind: string; text: string }>) {
-  return {
-    num_turns: 1,
-    usage: {
-      input_tokens: inputTokens,
-      output_tokens: 0,
-      cache_read_input_tokens: 0,
-      cache_creation_input_tokens: 0,
-      reasoning_tokens: 0,
-    },
-    tool_counts: Object.fromEntries(tools.map((tool) => [tool.kind, 1])),
-    turns: [
-      {
-        turn: 1,
-        usage: {
-          input_tokens: inputTokens,
-          output_tokens: 0,
-          cache_read_input_tokens: 0,
-          cache_creation_input_tokens: 0,
-          reasoning_tokens: 0,
-        },
-        tools,
-      },
-    ],
-  };
-}
-
-describe("PhaseLogCard usage", () => {
-  function renderCard(ui: ReactElement) {
-    return render(<MemoryRouter>{ui}</MemoryRouter>);
-  }
-
-  it("shows live token counts before the phase records spend", async () => {
-    const user = userEvent.setup();
-    const live = usageTelemetry(210, [{ kind: "bash", text: "git status" }]);
-    useLiveLogStreamMock.mockReturnValue({
-      ...idleLiveLogStream(vi.fn()),
-      telemetry: live,
-      usageSeries: [{ name: "Prompt", telemetry: live }],
-    });
-
-    renderCard(
-      <PhaseLogCard
-        phase={{ ...PHASE, status: "running", costCents: "0", totalTokens: "0", duration: "12s" }}
-        expanded={false}
-        organizationId="org-1"
-        canvasId="canvas-1"
-        stream={[
-          line({
-            id: "planner-agent",
-            componentName: "Agent",
-            component: "runnerClaudeCode",
-            executionId: "exec-1",
-            status: "running",
-          }),
-        ]}
-      />,
-    );
-
-    const showUsage = await screen.findByRole("button", { name: "Show usage" });
-    expect(screen.getByTestId("split-run-phase-duration-plan")).toHaveTextContent("210");
-    await user.click(showUsage);
-    expect(await screen.findByRole("heading", { name: "Plan usage" })).toBeInTheDocument();
-    expect(screen.getByText("1 turn · 1 tool call · 210 input")).toBeInTheDocument();
-    expect(screen.getByRole("status")).toHaveTextContent("Live. The run is not finished. New turns will appear here.");
-  });
-
-  it("opens a usage dialog from the cost and token text", async () => {
-    const user = userEvent.setup();
-    useLiveLogStreamMock.mockReturnValue({
-      ...idleLiveLogStream(vi.fn()),
-      telemetry: usageTelemetry(210, [
-        { kind: "bash", text: "git status" },
-        { kind: "read", text: "README.md" },
-      ]),
-    });
-
-    renderCard(
-      <PhaseLogCard
-        phase={{ ...PHASE, costCents: "45", totalTokens: "210" }}
-        expanded
-        organizationId="org-1"
-        canvasId="canvas-1"
-        stream={[
-          line({
-            id: "planner-agent",
-            componentName: "Agent - Plan for GH Issue",
-            component: "runnerClaudeCode",
-            executionId: "exec-1",
-          }),
-        ]}
-      />,
-    );
-
-    await user.click(screen.getByRole("button", { name: "Show usage" }));
-    expect(await screen.findByRole("heading", { name: "Plan usage" })).toBeInTheDocument();
-    expect(screen.getByText("1 turn · 2 tool calls · 210 input")).toBeInTheDocument();
-    expect(screen.queryByRole("heading", { name: "Agent - Plan for GH Issue" })).not.toBeInTheDocument();
-    await user.keyboard("{Escape}");
-    expect(screen.queryByRole("heading", { name: "Plan usage" })).not.toBeInTheDocument();
-  });
-
-  it("shows a separate chart for each agent", async () => {
-    const user = userEvent.setup();
-    const implementUsage = {
-      ...idleLiveLogStream(vi.fn()),
-      telemetry: usageTelemetry(180, [{ kind: "bash", text: "make test" }]),
-    };
-    const pullRequestUsage = {
-      ...idleLiveLogStream(vi.fn()),
-      telemetry: usageTelemetry(40, [{ kind: "read", text: "PR.md" }]),
-    };
-    useLiveLogStreamMock.mockImplementation((executionId: string) =>
-      executionId === "exec-pr" ? pullRequestUsage : implementUsage,
-    );
-
-    renderCard(
-      <PhaseLogCard
-        phase={{ ...PHASE, name: "Implement", costCents: "80", totalTokens: "220" }}
-        expanded
-        organizationId="org-1"
-        canvasId="canvas-1"
-        stream={[
-          line({
-            id: "impl-agent",
-            nodeId: "impl-agent",
-            componentName: "Implement the change",
-            component: "runnerClaudeCode",
-            executionId: "exec-impl",
-          }),
-          line({
-            id: "pr-agent",
-            nodeId: "pr-agent",
-            componentName: "Write the pull request",
-            component: "runnerClaudeCode",
-            executionId: "exec-pr",
-          }),
-        ]}
-      />,
-    );
-
-    await user.click(screen.getByRole("button", { name: "Show usage" }));
-    expect(await screen.findByRole("heading", { name: "Implement usage" })).toBeInTheDocument();
-    expect(screen.getByRole("heading", { name: "Implement the change" })).toBeInTheDocument();
-    expect(screen.getByRole("heading", { name: "Write the pull request" })).toBeInTheDocument();
-    expect(screen.getByText("1 turn · 1 tool call · 180 input")).toBeInTheDocument();
-    expect(screen.getByText("1 turn · 1 tool call · 40 input")).toBeInTheDocument();
-  });
-
-  it("shows a separate chart for each prompt on one agent", async () => {
-    const user = userEvent.setup();
-    useLiveLogStreamMock.mockReturnValue({
-      ...idleLiveLogStream(vi.fn()),
-      usageSeries: [
-        { name: "Implementation", telemetry: usageTelemetry(180, [{ kind: "bash", text: "make test" }]) },
-        {
-          name: "Generate PR title and description",
-          telemetry: usageTelemetry(40, [{ kind: "bash", text: "git log main..feature" }]),
-        },
-      ],
-    });
-
-    renderCard(
-      <PhaseLogCard
-        phase={{ ...PHASE, name: "Implement", costCents: "80", totalTokens: "220" }}
-        expanded
-        organizationId="org-1"
-        canvasId="canvas-1"
-        stream={[
-          line({
-            id: "impl-agent",
-            nodeId: "impl-agent",
-            componentName: "Implement the change",
-            component: "runnerClaudeCode",
-            executionId: "exec-impl",
-          }),
-        ]}
-      />,
-    );
-
-    await user.click(screen.getByRole("button", { name: "Show usage" }));
-    expect(await screen.findByRole("heading", { name: "Implement usage" })).toBeInTheDocument();
-    expect(screen.getByRole("heading", { name: "Implementation" })).toBeInTheDocument();
-    expect(screen.getByRole("heading", { name: "Generate PR title and description" })).toBeInTheDocument();
-    expect(screen.getByText("1 turn · 1 tool call · 180 input")).toBeInTheDocument();
-    expect(screen.getByText("1 turn · 1 tool call · 40 input")).toBeInTheDocument();
-  });
-});
-
-describe("PhaseLogCard node line", () => {
-  const NODE_STREAM: SplitRunStreamLine[] = [
-    line({
-      id: "planner-agent",
-      componentName: "Agent - Plan for GH Issue",
-      componentType: "Run Claude Code",
-      at: "12:24:02",
-      duration: "1m 20s",
-    }),
-  ];
-
-  it("shows the node name and puts status plus time on the far right", () => {
-    render(<PhaseLogCard phase={PHASE} expanded stream={NODE_STREAM} />);
-
-    const row = screen.getByTestId("split-run-stream-line-planner-agent");
-    expect(within(row).getByText("Agent - Plan for GH Issue")).toBeInTheDocument();
-    expect(within(row).queryByText("passed")).not.toBeInTheDocument();
-    expect(within(row).queryByText(">")).not.toBeInTheDocument();
-    expect(within(row).queryByText("12:24:02")).not.toBeInTheDocument();
-    expect(within(row).queryByText("Run Claude Code")).not.toBeInTheDocument();
-
-    const statusTime = within(row).getByTestId("split-run-stream-duration-planner-agent");
-    expect(statusTime).toHaveAccessibleName("Passed");
-    expect(statusTime).toHaveTextContent("✓");
-    expect(statusTime).toHaveTextContent("01:20");
-    expect(statusTime).not.toHaveTextContent("Passed");
-    expect(statusTime.className).not.toMatch(/bg-/);
-    expect(statusTime.parentElement?.className).toMatch(/ml-auto/);
-    expect(statusTime.className).toMatch(/text-right/);
-    expect(statusTime.className).toMatch(/font-mono/);
-    expect(statusTime.className).toMatch(/text-\[14px\]/);
-    expect(screen.getByTestId("split-run-stream-plan").className).toMatch(/font-mono/);
-  });
-
-  it("floats the produced artifact next to the duration", () => {
-    render(
-      <PhaseLogCard
-        phase={PHASE}
-        expanded
-        stream={[
-          line({
-            id: "planner-agent",
-            componentName: "Agent - Plan for GH Issue",
-            duration: "1m 20s",
-            artifact: {
-              id: "art-plan",
-              type: "TYPE_MARKDOWN",
-              data: { name: "PLAN.md", title: "PLAN.md" },
-            },
-          }),
-        ]}
-      />,
-    );
-
-    const row = screen.getByTestId("split-run-stream-line-planner-agent");
-    const artifact = within(row).getByRole("button", { name: "PLAN.md" });
-    const duration = within(row).getByTestId("split-run-stream-duration-planner-agent");
-
-    expect(artifact.parentElement).toBe(duration.parentElement);
-    expect(artifact.parentElement?.className).toMatch(/ml-auto/);
-    expect(artifact.compareDocumentPosition(duration) & Node.DOCUMENT_POSITION_FOLLOWING).toBeTruthy();
-  });
-
-  it("styles the node line as a table header without a caret", () => {
-    render(<PhaseLogCard phase={PHASE} expanded stream={NODE_STREAM} />);
-
-    const row = screen.getByTestId("split-run-stream-line-planner-agent");
-    expect(row.querySelector(".lucide-chevron-right")).toBeNull();
-    expect(row.className).toMatch(/\bbg-muted\b/);
-    expect(row.className).not.toMatch(/\bbg-background\b/);
-    expect(row.className).not.toMatch(/hover:bg-/);
-    expect(row.className).not.toMatch(/border-b/);
-    expect(row.className).not.toMatch(/-mx-2/);
-    const toggle = screen.getByTestId("split-run-node-toggle-planner-agent");
-    expect(toggle.className).toMatch(/gap-1\.5/);
-    expect(toggle.firstElementChild?.className).toMatch(/size-3/);
-    expect(screen.queryByTestId("split-run-node-indent")).not.toBeInTheDocument();
   });
 });
